+++ src/lib/services/validateForm.ts
@@ -1,99 +1,99 @@
-import { EventStore } from '$lib/stores/ObjectStore';
-import { get } from 'svelte/store';
+import { type DomainEvent } from '$lib/stores/ObjectStore';
+import { notificationStore } from '$lib/stores/notifications';
 
-interface FormError {
-	field: string;
-	message: string;
-}
-
-interface FormData {
-	target: Record<string, { name: string; required: boolean; value: string }>;
+export interface FormData {
+  target: Record<string, { name: string; required: boolean; value: string }>;
 }
 
 export function validateEmptyInput(formData: FormData): string[] {
-	const errObject = [];
-	for (const field in formData.target) {
-		if (formData.target[field].required && formData.target[field].value === '') {
-			errObject.push(formData.target[field].name);
-		}
-	}
-	return errObject;
+  if (!formData.target) {
+    return [];
+  }
+  const errObject = [];
+  for (const field in formData.target) {
+    if (formData.target[field].required && formData.target[field].value === '') {
+      errObject.push(formData.target[field].name);
+    }
+  }
+  return errObject;
 }
 
-export function validateEventData(): FormError[] {
-	const event = get(EventStore);
-	const formErrors: FormError[] = [];
+export const buildTimeStamp = (date: string, time: string): Date => {
+  return new Date(`${date}T${time}Z`);
+};
 
-	if (event.startdate && new Date(event.startdate) > new Date()) {
-		formErrors.push({
-			field: 'startdate',
-			message: 'The start date cannot be in the past',
-		});
-	}
-	if (event.enddate && event.startdate && new Date(event.enddate) < new Date(event.startdate)) {
-		formErrors.push({
-			field: 'enddate',
-			message: 'The end date cannot be before the start date',
-		});
-	}
-	if (
-		event.startdate &&
-		event.enddate &&
-		event.starttime &&
-		event.endtime &&
-		new Date(event.startdate) === new Date(event.enddate) &&
-		new Date(event.starttime) > new Date(event.endtime)
-	) {
-		formErrors.push({
-			field: 'endtime',
-			message: 'The end time cannot be before the start time',
-		});
-	}
-	if (
-		event.publishdate &&
-		event.startdate &&
-		new Date(event.publishdate) < new Date(event.startdate)
-	) {
-		formErrors.push({
-			field: 'publishdate',
-			message: 'The publish date cannot be before the start date',
-		});
-	}
-	if (
-		event.publishdate &&
-		event.startdate &&
-		event.publishtime &&
-		event.starttime &&
-		new Date(event.publishdate) === new Date(event.startdate) &&
-		new Date(event.publishtime) > new Date(event.starttime)
-	) {
-		formErrors.push({
-			field: 'publishtime',
-			message: 'The publish time cannot be before the start time',
-		});
-	}
-	if (
-		event.unpublishdate &&
-		event.publishdate &&
-		new Date(event.unpublishdate) < new Date(event.publishdate)
-	) {
-		formErrors.push({
-			field: 'unpublishdate',
-			message: 'The unpublish date cannot be before the publish date',
-		});
-	}
-	if (
-		event.unpublishdate &&
-		event.publishdate &&
-		event.unpublishtime &&
-		event.publishtime &&
-		new Date(event.unpublishdate) === new Date(event.publishdate) &&
-		new Date(event.unpublishtime) < new Date(event.publishtime)
-	) {
-		formErrors.push({
-			field: 'unpublishtime',
-			message: 'The unpublish time cannot be before the publish time',
-		});
-	}
-	return formErrors;
+export function validateEventData(event: DomainEvent): boolean {
+  if (!event) {
+    throw new Error('Event could be found');
+  }
+
+  const {
+    startdate,
+    starttime,
+    enddate,
+    endtime,
+    publishdate,
+    publishtime,
+    unpublishdate,
+    unpublishtime,
+  } = event;
+
+  const startDateTime = starttime && startdate && buildTimeStamp(startdate, starttime).getTime();
+  const endDateTime = enddate && endtime && buildTimeStamp(enddate, endtime).getTime();
+  const publishDateTime =
+    publishdate && publishtime && buildTimeStamp(publishdate, publishtime).getTime();
+  const unpublishDateTime =
+    unpublishdate && unpublishtime && buildTimeStamp(unpublishdate, unpublishtime).getTime();
+
+  let hasError = false;
+
+  if (!startdate) {
+    notificationStore.addToast('error', 'Mandatory start date is empty');
+    hasError = true;
+  } else if (!starttime) {
+    notificationStore.addToast('error', 'Mandatory start time is empty');
+    hasError = true;
+  }
+
+  if (!enddate) {
+    notificationStore.addToast('error', 'Mandatory end date is empty');
+    hasError = true;
+  } else if (!endtime) {
+    notificationStore.addToast('error', 'Mandatory end time is empty');
+    hasError = true;
+  }
+
+  if (new Date(startdate) < new Date()) {
+    notificationStore.addToast('error', 'The start date cannot be in the past');
+    hasError = true;
+  }
+
+  if (startDateTime && endDateTime) {
+    if (endDateTime < startDateTime) {
+      notificationStore.addToast('error', 'The event end cannot be before the event start');
+      hasError = true;
+    } else if (endDateTime === startDateTime) {
+      notificationStore.addToast('warning', 'Both start and end date-time are the same');
+      hasError = true;
+    }
+  }
+
+  if (startDateTime && publishDateTime && startDateTime < publishDateTime) {
+    notificationStore.addToast('error', 'The publish date cannot be later than the start date');
+    hasError = true;
+  }
+
+  if (unpublishDateTime && publishDateTime) {
+    if (unpublishDateTime < publishDateTime) {
+      notificationStore.addToast(
+        'error',
+        'The unpublish date cannot be earlier than the publish date'
+      );
+      hasError = true;
+    } else if (unpublishDateTime === publishDateTime) {
+      notificationStore.addToast('warning', 'Both publish and unpublish date-time are the same.');
+      hasError = true;
+    }
+  }
+  return hasError;
 }