+++ src/routes/(admin)/admin/newsadmin/create/+page.svelte
@@ -4,15 +4,29 @@
 	import { addDoc } from 'firebase/firestore';
 	import { newsColRef } from '$lib/firebase/firebaseConfig';
 
-	import type { News } from '$lib/stores/ObjectStore';
+	import { type News, EditModeStore } from '$lib/stores/ObjectStore';;
+	import {selectedImage, imageExists, existingImageUrl} from '$lib/stores/ImageSelectionStore';
+	import { newsFormService } from '$lib/services/NewsFormService';
 
 	import NewsForm from '$lib/components/NewsForm.svelte';
+	import { notificationStore } from '$lib/stores/notifications';
+	import ToastContainer from '$lib/components/ToastContainer.svelte';
 
 	const saveNewItem = async (newNewsItem: News) => {
+		
 		try {
-			await addDoc(newsColRef, newNewsItem);
+			const thisNews: News | undefined = await newsFormService(
+				newNewsItem,
+				$selectedImage,
+				$imageExists,
+				$existingImageUrl
+			);
+			await addDoc(newsColRef, thisNews);
 			goto('/admin/newsadmin');
+			EditModeStore.set('');
+			notificationStore.addToast('success', 'News article created successfully!');
 		} catch (error) {
+			notificationStore.addToast('error', 'Failed to create news article. Please try again.');
 			console.error('Error writing document:', error);
 		}
 	};
@@ -21,3 +35,5 @@
 <div>
 	<NewsForm onCreateNew={saveNewItem} />
 </div>
+
+<ToastContainer />